Refactor - Moved profile classes
