+++ common/src/main/java/com/scottlogic/deg/common/profile/Profile.java
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package com.scottlogic.deg.generator.profile;
+package com.scottlogic.deg.common.profile;
 
 import com.scottlogic.deg.common.profile.fields.Field;
 import com.scottlogic.deg.common.profile.fields.Fields;
+import com.scottlogic.deg.common.profile.rules.Rule;
 
 import java.util.Collection;
 import java.util.List;
