+++ generator/src/main/java/com/scottlogic/deg/generator/fieldspecs/relations/EqualToRelation.java
@@ -21,7 +21,7 @@ import com.scottlogic.deg.generator.fieldspecs.FieldSpec;
 import com.scottlogic.deg.generator.fieldspecs.FieldSpecFactory;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.DistributedList;
 import com.scottlogic.deg.generator.generation.databags.DataBagValue;
-import com.scottlogic.deg.generator.profile.constraints.Constraint;
+import com.scottlogic.deg.common.profile.rules.constraints.Constraint;
 
 public class EqualToRelation implements FieldSpecRelations {
     private final Field main;
