+++ generator/src/main/java/com/scottlogic/deg/generator/fieldspecs/relations/InMapIndexRelation.java
@@ -21,7 +21,7 @@ import com.scottlogic.deg.generator.fieldspecs.FieldSpec;
 import com.scottlogic.deg.generator.fieldspecs.FieldSpecFactory;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.DistributedList;
 import com.scottlogic.deg.generator.generation.databags.DataBagValue;
-import com.scottlogic.deg.generator.profile.constraints.Constraint;
+import com.scottlogic.deg.common.profile.rules.constraints.Constraint;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
