+++ generator/src/main/java/com/scottlogic/deg/generator/generation/DataGenerator.java
@@ -16,7 +16,7 @@
 
 package com.scottlogic.deg.generator.generation;
 
-import com.scottlogic.deg.generator.profile.Profile;
+import com.scottlogic.deg.common.profile.Profile;
 import com.scottlogic.deg.common.output.GeneratedObject;
 
 import java.util.stream.Stream;
