+++ generator/src/main/java/com/scottlogic/deg/generator/generation/DecisionTreeDataGenerator.java
@@ -19,7 +19,7 @@ package com.scottlogic.deg.generator.generation;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import com.scottlogic.deg.common.output.GeneratedObject;
-import com.scottlogic.deg.generator.profile.Profile;
+import com.scottlogic.deg.common.profile.Profile;
 import com.scottlogic.deg.generator.config.detail.VisualiserLevel;
 import com.scottlogic.deg.generator.decisiontree.DecisionTree;
 import com.scottlogic.deg.generator.decisiontree.DecisionTreeFactory;
