+++ generator/src/main/java/com/scottlogic/deg/generator/inputs/validation/UniquenessValidator.java
@@ -3,7 +3,7 @@ package com.scottlogic.deg.generator.inputs.validation;
 import com.google.inject.Inject;
 import com.scottlogic.deg.common.ValidationException;
 import com.scottlogic.deg.common.profile.fields.Field;
-import com.scottlogic.deg.generator.profile.Profile;
+import com.scottlogic.deg.common.profile.Profile;
 import com.scottlogic.deg.generator.generation.GenerationConfigSource;
 
 import static com.scottlogic.deg.generator.config.detail.CombinationStrategyType.MINIMAL;
