+++ generator/src/main/java/com/scottlogic/deg/generator/profile/constraints/atomic/AtomicConstraint.java
@@ -18,7 +18,7 @@ package com.scottlogic.deg.generator.profile.constraints.atomic;
 
 import com.scottlogic.deg.common.profile.fields.Field;
 import com.scottlogic.deg.generator.fieldspecs.FieldSpec;
-import com.scottlogic.deg.generator.profile.constraints.Constraint;
+import com.scottlogic.deg.common.profile.rules.constraints.Constraint;
 
 public interface AtomicConstraint extends Constraint {
