+++ generator/src/main/java/com/scottlogic/deg/generator/profile/constraints/grammatical/OrConstraint.java
@@ -16,7 +16,7 @@
 
 package com.scottlogic.deg.generator.profile.constraints.grammatical;
 
-import com.scottlogic.deg.generator.profile.constraints.Constraint;
+import com.scottlogic.deg.common.profile.rules.constraints.Constraint;
 
 import java.util.Arrays;
 import java.util.Collection;
