+++ generator/src/main/java/com/scottlogic/deg/generator/violations/filters/ConstraintTypeViolationFilter.java
@@ -16,7 +16,7 @@
 
 package com.scottlogic.deg.generator.violations.filters;
 
-import com.scottlogic.deg.generator.profile.constraints.Constraint;
+import com.scottlogic.deg.common.profile.rules.constraints.Constraint;
 
 /**
  * Violation filter which filters on teh type of the constraint.
