+++ generator/src/test/java/com/scottlogic/deg/generator/builders/OrBuilder.java
@@ -16,7 +16,7 @@
 
 package com.scottlogic.deg.generator.builders;
 
-import com.scottlogic.deg.generator.profile.constraints.Constraint;
+import com.scottlogic.deg.common.profile.rules.constraints.Constraint;
 import com.scottlogic.deg.generator.profile.constraints.grammatical.OrConstraint;
 
 import java.util.List;
