+++ generator/src/test/java/com/scottlogic/deg/generator/profile/constraints/grammatical/NotConstraintTests.java
@@ -17,7 +17,7 @@
 package com.scottlogic.deg.generator.profile.constraints.grammatical;
 
 import com.scottlogic.deg.common.profile.fields.Field;
-import com.scottlogic.deg.generator.profile.constraints.Constraint;
+import com.scottlogic.deg.common.profile.rules.constraints.Constraint;
 import com.scottlogic.deg.generator.profile.constraints.atomic.IsInSetConstraint;
 import com.scottlogic.deg.generator.profile.constraints.atomic.IsNullConstraint;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.DistributedList;
