+++ orchestrator/src/main/java/com/scottlogic/deg/orchestrator/violate/ViolateOutputValidator.java
@@ -18,7 +18,7 @@ package com.scottlogic.deg.orchestrator.violate;
 
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
-import com.scottlogic.deg.generator.profile.Profile;
+import com.scottlogic.deg.common.profile.Profile;
 import com.scottlogic.deg.output.OutputPath;
 import com.scottlogic.deg.output.outputtarget.OutputTargetValidationException;
 import com.scottlogic.deg.common.util.FileUtils;
