+++ orchestrator/src/main/java/com/scottlogic/deg/orchestrator/violate/ViolatedProfile.java
@@ -16,9 +16,9 @@
 
 package com.scottlogic.deg.orchestrator.violate;
 
-import com.scottlogic.deg.generator.profile.Profile;
+import com.scottlogic.deg.common.profile.Profile;
 import com.scottlogic.deg.common.profile.fields.Fields;
-import com.scottlogic.deg.generator.profile.Rule;
+import com.scottlogic.deg.common.profile.rules.Rule;
 
 import java.util.Collection;
