+++ orchestrator/src/test/java/com/scottlogic/deg/orchestrator/violate/ViolateOutputValidatorTests.java
@@ -16,8 +16,8 @@
 
 package com.scottlogic.deg.orchestrator.violate;
 
-import com.scottlogic.deg.generator.profile.Profile;
-import com.scottlogic.deg.generator.profile.Rule;
+import com.scottlogic.deg.common.profile.Profile;
+import com.scottlogic.deg.common.profile.rules.Rule;
 import com.scottlogic.deg.common.util.FileUtils;
 import com.scottlogic.deg.output.OutputPath;
 import com.scottlogic.deg.output.outputtarget.OutputTargetValidationException;
