+++ orchestrator/src/test/java/com/scottlogic/deg/orchestrator/violate/violator/ProfileViolatorTests.java
@@ -19,10 +19,10 @@ package com.scottlogic.deg.orchestrator.violate.violator;
 import com.scottlogic.deg.common.profile.fields.Field;
 import com.scottlogic.deg.common.profile.HelixNumber;
 import com.scottlogic.deg.common.profile.fields.Fields;
-import com.scottlogic.deg.generator.profile.Profile;
-import com.scottlogic.deg.generator.profile.Rule;
-import com.scottlogic.deg.generator.profile.RuleInformation;
-import com.scottlogic.deg.generator.profile.constraints.Constraint;
+import com.scottlogic.deg.common.profile.Profile;
+import com.scottlogic.deg.common.profile.rules.Rule;
+import com.scottlogic.deg.common.profile.rules.RuleInformation;
+import com.scottlogic.deg.common.profile.rules.constraints.Constraint;
 import com.scottlogic.deg.generator.profile.constraints.atomic.IsGreaterThanConstantConstraint;
 import com.scottlogic.deg.generator.profile.constraints.atomic.IsLessThanConstantConstraint;
 import com.scottlogic.deg.orchestrator.violate.ViolatedProfile;
