+++ profile/src/main/java/com/scottlogic/deg/profile/reader/AtomicConstraintReader.java
@@ -23,7 +23,7 @@ import com.scottlogic.deg.common.profile.fields.Fields;
 import com.scottlogic.deg.common.util.NumberUtils;
 import com.scottlogic.deg.generator.fieldspecs.relations.InMapRelation;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.DistributedList;
-import com.scottlogic.deg.generator.profile.constraints.Constraint;
+import com.scottlogic.deg.common.profile.rules.constraints.Constraint;
 import com.scottlogic.deg.generator.profile.constraints.atomic.*;
 import com.scottlogic.deg.profile.dtos.constraints.*;
 import org.jetbrains.annotations.Nullable;
